Print out the goroutine id
This depends on 'loggoroutine' build tag
