+++ formatter/hook/additional_empty.go
@@ -1,0 +1,8 @@
+//go:build !loggoroutine
+
+package hook
+
+import log "github.com/sirupsen/logrus"
+
+func additionalEntries(_ *log.Entry) {
+}
