+++ formatter/hook/hook_test.go
@@ -1,4 +1,4 @@
-package formatter
+package hook
 
 import (
 	"testing"
