+++ formatter/hook/keys.go
@@ -1,0 +1,6 @@
+package hook
+
+const (
+	EntryKeySource      = "source"
+	EntryKeyGoroutineID = "goroutine_id"
+)
